feat: default to gcoredump key extraction on macOS
